// src/shared/enums/exceptions.ts
export enum EAPIErrorType {
  TECHNICAL_ERROR = 'TECHNICAL_ERROR',
  INVALID_PARAMETER = 'INVALID_PARAMETER',
  RESOURCE_NOT_FOUND = 'RESOURCE_NOT_FOUND',
  CODE_EXPIRED = 'CODE_EXPIRED',
  UNAUTHORIZED = 'UNAUTHORIZED',
  INVALID_FILE_FORMAT = 'INVALID_FILE_FORMAT',
  INVALID_FILE_CONTENT = 'INVALID_FILE_CONTENT',
  INVALID_FILE_ENCODING = 'INVALID_FILE_ENCODING',
  NO_FILE_UPLOADED = 'NO_FILE_UPLOADED',
  ONLY_CSV_AND_EXCEL_FILES_ARE_ALLOWED = 'ONLY_CSV_AND_EXCEL_FILES_ARE_ALLOWED',
  CREATE_ENTITY_ERROR = 'CREATE_ENTITY_ERROR',
  UPDATE_ENTITY_ERROR = 'UPDATE_ENTITY_ERROR',
  DELETE_ENTITY_ERROR = 'DELETE_ENTITY_ERROR',
  FETCH_ENTITIES_ERROR = 'FETCH_ENTITIES_ERROR',
  FETCH_ENTITY_BY_ID_ERROR = 'FETCH_ENTITY_BY_ID_ERROR',
  FETCH_ENTITY_BY_FIELD_ERROR = 'FETCH_ENTITY_BY_FIELD_ERROR',
  AUTH_TOKEN_ERROR = 'AUTH_TOKEN_ERROR',
  FILE_VALIDATION_ERROR = 'FILE_VALIDATION_ERROR',
}